refactor(ui): add dbflux_components::modals and move ModalShell (#102)
Create modals/ module in dbflux_components with shell.rs (ModalShell,
ModalVariant) rewritten to use crate:: intra-crate paths. Register pub mod
modals in lib.rs. Replace overlays/modals/shell.rs with a pub-use shim so
existing call-sites in the five pure modals continue to resolve.
