+++ crates/dbflux_components/src/modals/mod.rs
@@ -1,0 +1,3 @@
+pub mod shell;
+
+pub use shell::{ModalShell, ModalVariant};
